sub: add subrandr renderer support
subrandr is a subtitle rendering library which aims to render
SRV3 (YouTube) subtitles and WebVTT subtitles accurately.

Currently in mpv WebVTT subs are rendered via ffmpeg conversion to ASS
which throws away a lot of the style and completely disregards
the WebVTT non-region-cue positioning algorithm. Furthermore if
one wants to render some more complex SRV3 subtitles one has to
resort to external converters since it's not even supported by ffmpeg.

However subrandr is able to render SRV3 subtitles natively with support
for the most commonly used features. It can render ruby text without
relying on font metrics during conversion which is obviously fragile,
it can perform correct scaling using the exact calculations used by
YouTube instead of making up ASS approximations. Similarly it follows
the WebVTT spec for the features of WebVTT that it supports.

It's not perfect of course and the library is still in "early" stages
of development (even though it's been 7 months). Notable unimplemented
things currently are:
- No DirectWrite font provider (no fonts will be found on Windows)
- No CoreText font provider (no fonts will be found on MacOS)
- Line-breaking is very naive and does unnecessary reshaping.
  Lines are only broken on whitespace instead of following the
  Unicode line breaking algorithm.
- Unicode bidirectional algorithm is not followed.
- Any form of vertical text is unsupported.